requirements.txt
flask
flask_cors
requests
psycopg2-binary